Move factor methods to `type-factor.c`
